+++ packages/core/cli/src/commands/start.js
@@ -13,11 +13,18 @@ module.exports = (cli) => {
   cli
     .command('start')
     .option('-p, --port [port]')
+    .option('-d, --daemon')
     .option('--db-sync')
     .allowUnknownOption()
     .action(async (opts) => {
       if (opts.port) {
         process.env.APP_PORT = opts.port;
+      }
+      if (process.platform === 'win32') {
+        console.log(
+          chalk.yellow(`It is not supported on win platform, please use \`${chalk.green('yarn dev')}\` instead`),
+        );
+        return;
       }
       if (process.argv.includes('-h') || process.argv.includes('--help')) {
         promptForTs();
@@ -42,6 +49,15 @@ module.exports = (cli) => {
       if (opts.dbSync) {
         await run('node', [`./packages/${APP_PACKAGE_ROOT}/server/lib/index.js`, 'db:sync']);
       }
-      run('pm2-runtime', ['start', `packages/${APP_PACKAGE_ROOT}/server/lib/index.js`, '--', ...process.argv.slice(2)]);
+      if (opts.daemon) {
+        run('pm2', ['start', `packages/${APP_PACKAGE_ROOT}/server/lib/index.js`, '--', ...process.argv.slice(2)]);
+      } else {
+        run('pm2-runtime', [
+          'start',
+          `packages/${APP_PACKAGE_ROOT}/server/lib/index.js`,
+          '--',
+          ...process.argv.slice(2),
+        ]);
+      }
     });
 };